feat(contrib): implement a to and write function
The `to()` and `write()` function corresponds directly to the `from()`
and `select()` functions. The `to()` function will write data using the
client with the `_measurement`, `_field`, and `_value` columns while
using the group key for tags. The `write()` function will write to a
single measurement and it will write all non-group key columns as fields
and the tags as tags.

This also introduces another dependency called the `PointsWriter`. The
default is to use an http points writer, but it can be overridden to
write directly to the database when embedded into influxdb while using
the same code.